refac: relocate auth related code to auth package
