+++ src/main/java/nextstep/SubwayApplication.java
@@ -2,6 +2,7 @@ package nextstep;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
 public class SubwayApplication {
